requirements.txt
Flask
flask-jwt-extended
flask-sqlalchemy
flask-cors
pymysql
python-dotenv
bcrypt